+++ lib/kanza.dart
@@ -1,7 +1,6 @@
 library kanza;
 
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+export 'src/kanza.dart';
+export 'src/kanza_builder.dart';
+export 'src/kanza_listener.dart';
+export 'src/kanza_provider.dart';